pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-cli"
version = "0.1.0"
description = "A simple, opinionated local RAG CLI tool."
authors = [
    {name = "Utsav Mehrotra", email = "utsavmehrotra23@gmail.com"},
]
dependencies = [
    "openai",
    "faiss-cpu",
    "numpy",
    "pypdf",
    "typer[all]",
    "pyyaml",
    "rich"
]
requires-python = ">=3.9"

[project.scripts]
rag = "rag.cli:app"
